* commit '35f7f957828c5ae3d5af24496d3371be0cd186f4':
  基金详
  .我的基金，更多设置隐私关闭参与基金排行，仍可以修改开放持仓。
  Add version code ;

+++ src/com/dkhs/portfolio/ui/PositionAdjustActivity.java
@@ -632,7 +632,7 @@ public class PositionAdjustActivity extends ModelAcitivity implements IDutyNotif
             SubmitSymbol symbol = new SubmitSymbol();
             symbol.setSymbol(stock.getStockId());
             // symbol.setPercent(stock.getPercent() / 100);
-            symbol.setPercent(stock.getPercent());
+            symbol.setPercent((int)stock.getPercent());
             // System.out.println("symbols stock id:" + symbol.getSymbol() + " value:" + symbol.getPercent());
             symbols.add(symbol);
         }
